Make LogCompactionSuite a new suite
